Set correct signature when compressed or different
Before the algorithm did not cope with compressed phars and/or changing
of the signature.
It also required that the last file was ended with "\n".

Now it is possible to change the signature in the save call (as the
previous signature is killed) and compressed phars are also handled
correctly (as the compressed content now may contain "\n").