fix signature begin determination
previously the begin of the signature was determined based on the stub-
length.

NOTE: determining the signature begin based on manifest parser end position
      can lead to overwriting the last byte of the last file in the phar
      with the first byte of the hash.

this ignored the size of the manifest.

fix is to take the end of the manifest which is known due to it's size
and add the compressed file sizes to it instead of adding this to the
internal position of the parser that runs only to detect the compressed
file sizes. authoritative is the manifest size.